requirements.txt
# Core Dependencies
streamlit==1.28.2
pandas==2.1.3
numpy==1.24.3
yfinance==0.2.33
plotly==5.18.0

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0
xgboost==2.0.2
statsmodels==0.14.0

# News Scraping and Sentiment Analysis
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
newsapi-python==0.2.7
vaderSentiment==3.3.2
textblob==0.17.1

# Technical Analysis
ta==0.11.0
pandas-ta==0.3.14b0

# Database
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
joblib==1.3.2
tqdm==4.66.1

# Data Validation
pydantic==2.5.2

# Caching
diskcache==5.6.3